+++ cupyx/distributed/array/_index_arith.py
